+++ upproject/src/main/java/com/example/upproject/ResistActivity.java
@@ -5,14 +5,13 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.os.Handler;
 import android.os.Message;
-import android.util.Log;
 import android.view.View;
 import android.view.Window;
 import android.view.WindowManager;
 import android.widget.Button;
 import android.widget.EditText;
-import android.widget.ImageButton;
-import android.widget.Toast;
+
+import com.example.upproject.ui.main.MainActivity;
 
 /**
  * Created by cjs on 2016/3/14.
