+++ upproject/src/main/java/com/example/upproject/sql/SqilteHelper.java
@@ -1,10 +1,11 @@
-package com.example.upproject;
+package com.example.upproject.sql;
 
 import android.content.Context;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteDatabase.CursorFactory;
 import android.database.sqlite.SQLiteOpenHelper;
 import android.util.Log;
+
 /**
  * Created by myalien on 2016/7/17.
  */
