+++ upproject/src/main/java/com/example/upproject/sql/UserInfo.java
@@ -1,4 +1,4 @@
-package com.example.upproject;
+package com.example.upproject.sql;
 
 import java.io.Serializable;
