+++ upproject/src/main/java/com/example/upproject/ui/main/ui_ietm_list/FirstListItem.java
@@ -1,11 +1,11 @@
-package com.example.upproject;
+package com.example.upproject.ui.main.ui_ietm_list;
 
 import android.app.Activity;
 import android.content.Intent;
+import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.os.Handler;
 import android.os.Message;
-import android.util.Log;
 import android.view.View;
 import android.view.Window;
 import android.view.WindowManager;
@@ -13,12 +13,17 @@ import android.widget.Button;
 import android.widget.CompoundButton;
 import android.widget.EditText;
 import android.widget.TextView;
+import android.widget.Toast;
 import android.widget.ToggleButton;
+
+import com.example.upproject.ConnectServerWithSocket;
+import com.example.upproject.R;
+import com.example.upproject.Tolerate;
 
 /**
  * Created by cjs on 2016/3/13.
  */
-public class Listitem4 extends Activity {
+public class FirstListItem extends Activity {
     private ToggleButton toggleButton;
     private TextView state;
     private EditText newpower;
@@ -33,23 +38,28 @@ public class Listitem4 extends Activity {
     private String newtime;
     private Button refresh;
     private Button more;
-
+    private String refreach_power="0";//刷新后的功率
+    private String refreach_rongren;//刷新后的容忍度
+    private String refreach_state;//刷新后的时间
+    private String refreach_time;//刷新后的时间
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.listitem4);
+        setContentView(R.layout.firstitem);
+        change = (Button) findViewById(R.id.change);
+        refresh= (Button) findViewById(R.id.refresh);
         init();
         more.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                Intent intent=new Intent(Listitem4.this,Tolerate4.class);
+                Intent intent=new Intent(FirstListItem.this,Tolerate.class);
                 startActivity(intent);
             }
         });
         refresh.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                sendmsg="4 car 20 ";
+                sendmsg="4 air 20 ";
                 ConnectServerWithSocket text1 = new ConnectServerWithSocket();
                 text1.setStr(sendmsg);
                 text1.start();
@@ -59,7 +69,22 @@ public class Listitem4 extends Activity {
                     e.printStackTrace();
                 }
                 replay = text1.getReplayfromserver();
-                Log.e("refresh4",replay);
+                if(replay!=null)
+                {
+                    refreach_power=replay.substring(0,1);//刷新得到的功率
+                    refreach_time=replay.substring(3,4);//刷新后的单次运行时间
+                    refreach_rongren=replay.substring(6,7);//刷新后的容忍度
+                    refreach_time=replay.substring(9,10);//刷新后的时间
+
+                    Toast.makeText(FirstListItem.this, "目前的功率为: " + refreach_power + "/n目前的单次运行时间为： " +
+                            refreach_time + "/n目前的容忍度为： " + refreach_rongren, Toast.LENGTH_SHORT).show();
+
+                }
+
+
+                Message message = handler.obtainMessage();
+                message.what = 5;
+                handler.sendMessage(message);
             }
         });
 
@@ -70,7 +95,7 @@ public class Listitem4 extends Activity {
                 tv_power = (TextView) findViewById(R.id.gonglv);
                 power = newpower.getText().toString();
 
-                sendmsg="4 car 00 "+power;
+                sendmsg="4 air 00 "+power;
                 ConnectServerWithSocket text1 = new ConnectServerWithSocket();
                 text1.setStr(sendmsg);
                 text1.start();
@@ -94,7 +119,7 @@ public class Listitem4 extends Activity {
             @Override
             public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
                 if (isChecked) {
-                    sendmsg="4 car 11";
+                    sendmsg="4 air 11";
                     ConnectServerWithSocket text = new ConnectServerWithSocket();
                     text.setStr(sendmsg);
                     text.start();
@@ -108,7 +133,7 @@ public class Listitem4 extends Activity {
                     message.what = 1;
                     handler.sendMessage(message);
                 } else {
-                    sendmsg="4 car 10";
+                    sendmsg="4 air 10";
                     ConnectServerWithSocket text = new ConnectServerWithSocket();
                     text.setStr(sendmsg);
                     text.start();
@@ -136,7 +161,9 @@ public class Listitem4 extends Activity {
         window.setFlags(WindowManager.LayoutParams.FLAG_TRANSLUCENT_NAVIGATION, WindowManager.LayoutParams
                 .FLAG_TRANSLUCENT_NAVIGATION);
         newpower = (EditText) findViewById(R.id.newpower);
+
         //设置用电器单次工作时间
+        //rongrendu= (TextView) findViewById(R.id.rongren);
         tv_worktime= (TextView) findViewById(R.id.tv_worktime);
         et_worktime= (EditText) findViewById(R.id.et_worktime);
         btn_change2= (Button) findViewById(R.id.changge2);
@@ -146,7 +173,7 @@ public class Listitem4 extends Activity {
             public void onClick(View v) {
                 newtime=et_worktime.getText().toString();
 
-                sendmsg="4 car 01 "+newtime;
+                sendmsg="4 air 01 "+newtime;
                 ConnectServerWithSocket text1 = new ConnectServerWithSocket();
                 text1.setStr(sendmsg);
                 text1.start();
@@ -162,27 +189,40 @@ public class Listitem4 extends Activity {
             }
         });
         //改变用电器功率
-        change = (Button) findViewById(R.id.change);
-        refresh= (Button) findViewById(R.id.refresh);
+
     }
 
     private Handler handler = new Handler() {
         @Override
         public void handleMessage(Message msg) {
+            SharedPreferences read=getSharedPreferences("DATA_TABLE",MODE_WORLD_READABLE);
+            String value;
             super.handleMessage(msg);
             switch (msg.what) {
                 case 1:
-                    state.setText("运行状态：开启");
+                    value=read.getString("STATE1","");
+                    if (value.equals("1")){
+                    state.setText("运行状态：开启");}
                     break;
                 case 2:
                     state.setText("运行状态：关闭");
                     break;
                 case 3:
-                    tv_power.setText("功率大小：" + power + "kw");
-
+                    value=read.getString("POWER1","");
+                    tv_power.setText("功率大小：" + value + "kw");//改动部分
                     break;
                 case 4:
-                    tv_worktime.setText("单次运行时间："+newtime+"小时");
+                    value=read.getString("TIME1","");
+                    tv_worktime.setText("单次运行时间："+value+"小时");
+                    break;
+                case 5:
+                    String value1=read.getString("POWER1","");
+                    String value2=read.getString("TIME1","");
+                    String value3=read.getString("STATE1","");
+                    String value4=read.getString("RONGREN1","");//得到容忍度，还没进行设置
+                    tv_power.setText("功率大小：" + value1 + "kw");
+                    tv_worktime.setText("单次运行时间："+value2+"小时");
+                    break;
             }
         }
     };
