+++ upproject/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
      android:orientation="vertical"
-    tools:context="com.example.upproject.MainActivity">
+    tools:context=".ui.main.MainActivity">
 
     <android.support.v4.view.ViewPager
         android:id="@+id/viewpager"
